+++ rust/operator-binary/src/extensions.rs
@@ -116,11 +116,12 @@ mod tests {
                                 vec![],
                                 None
                             ),
-                            oidc: oidc::v1alpha1::ClientAuthenticationOptions {
+                            oidc: crate::authentication::oidc::DruidClientAuthenticationOptions {
                                 client_credentials_secret_ref: "".to_string(),
                                 extra_scopes: vec![],
-                                client_authentication_method: Default::default(),
-                                product_specific_fields: ()
+                                product_specific_fields: oidc::v1alpha1::ClientAuthenticationMethodOption {
+                                    client_authentication_method: oidc::v1alpha1::ClientAuthenticationMethod::default(),
+                                },
                             }
                         }]
                     })