+++ rust/operator-binary/src/listener.rs
@@ -60,7 +60,7 @@ pub fn build_group_listener(
             .name(listener_group_name)
             .ownerreference_from_resource(druid, None, Some(true))
             .context(ObjectMissingMetadataForOwnerRefSnafu)?
-            .with_recommended_labels(object_labels)
+            .with_recommended_labels(&object_labels)
             .context(BuildObjectMetaSnafu)?
             .build(),
         spec: listener::v1alpha1::ListenerSpec {