+++ CHANGELOG.md
@@ -6,9 +6,12 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- BREAKING: `configOverrides` now only accepts the supported config file names `runtime.properties`, `jvm.config` and `security.properties`. Previously arbitrary keys were silently accepted but ignored ([#XXX]).
+- Bump `stackable-operator` to 0.110.0 and `kube` to 3.1.0 ([#XXX]).
 - Document Helm deployed RBAC permissions and remove unnecessary permissions ([#810]).
 
-- [#810]: https://github.com/stackabletech/druid-operator/pull/810
+[#XXX]: https://github.com/stackabletech/druid-operator/pull/XXX
+[#810]: https://github.com/stackabletech/druid-operator/pull/810
 
 ## [26.3.0] - 2026-03-16